chore: configure eslint and fix build issues
