+++ components/products/product-inventory-history.tsx
@@ -1,0 +1,13 @@
+"use client"
+
+interface ProductInventoryHistoryProps {
+  transactions: any[]
+}
+
+export function ProductInventoryHistory({ transactions }: ProductInventoryHistoryProps) {
+  return (
+    <div>
+      <p>Lịch sử tồn kho đang được phát triển.</p>
+    </div>
+  )
+}
