+++ package.json
@@ -70,6 +70,7 @@
     "postcss": "^8.5",
     "tailwindcss": "^4.1.9",
     "tw-animate-css": "1.3.3",
-    "typescript": "^5"
+    "typescript": "^5",
+    "eslint": "^9.0.0"
   }
 }
